move debounce options into useDebounceOptions hook
