+++ src/views/debounceView/useDebounceOptions.ts
@@ -1,4 +1,5 @@
 import { ChangeEventHandler, useState } from "react";
+import { VisualizationOptions } from "../../layouts/visualizationLayout/VisualizationLayout";
 
 export default function useDebounceOptions() {
   const [isLeading, setIsLeading] = useState<"true" | "false">("false");
@@ -19,14 +20,84 @@ export default function useDebounceOptions() {
     setWait(e.target.value);
   };
 
+  const options: VisualizationOptions[] = [
+    {
+      type: "number",
+      inputGroupProps: {
+        label: "Wait(ms)",
+        showLabel: true,
+        value: wait,
+        onChange: handleWait,
+        name: "wait",
+        min: 0,
+        step: 100,
+        tooltipProps: {
+          content:
+            "This is the waiting period. Any attempt to call the debounced function during the waiting period will reset it.",
+        },
+      },
+    },
+    {
+      type: "dropdown",
+      dropdownProps: {
+        label: "Leading",
+        showLabel: true,
+        name: "leading",
+        value: isLeading,
+        onChange: handleLeading,
+        options: [
+          { value: "true", children: "true" },
+          { value: "false", children: "false" },
+        ],
+        tooltipProps: {
+          content:
+            "If leading is set to true, the debounced function will execute once before initiating the waiting period.",
+        },
+      },
+    },
+    {
+      type: "dropdown",
+      dropdownProps: {
+        label: "Trailing",
+        showLabel: true,
+        name: "trailing",
+        value: isTrailing,
+        onChange: handleTrailing,
+        options: [
+          { value: "true", children: "true" },
+          { value: "false", children: "false" },
+        ],
+        tooltipProps: {
+          content:
+            "If trailing is set to true, the debounced function will execute after the waiting period is over.",
+        },
+      },
+    },
+    {
+      type: "dropdown",
+      dropdownProps: {
+        label: "Cancellable",
+        showLabel: true,
+        name: "cancellable",
+        value: isCancellable,
+        onChange: handleCancellable,
+        options: [
+          { value: "true", children: "true" },
+          { value: "false", children: "false" },
+        ],
+        tooltipProps: {
+          content:
+            "This allows you to cancel the debounced function if it hasn't executed yet.",
+        },
+      },
+    },
+  ];
+
   return {
     isLeading,
-    handleLeading,
     isTrailing,
-    handleTrailing,
     isCancellable,
-    handleCancellable,
     wait,
-    handleWait,
+    options,
   };
 }
